Refactor and cleanup notification components
