+++ src/components/notification/constants.ts
@@ -1,0 +1,1 @@
+export const NOTIFICATION_MODAL_ID = "notification_modal";
