+++ src/components/notification/index.ts
@@ -1,7 +1,6 @@
 export { default as NotificationButton } from "./NotificationButton";
-export type { INotificationButtonProps } from "./NotificationButton";
-export { NotificationButtonType } from "./NotificationButton";
-
 export { default as NotificationComponent } from "./NotificationComponent";
-export type { NotificationData } from "./NotificationComponent";
-export { NOTIFICATION_MODAL_ID } from "./NotificationComponent";
+
+export type { NotificationData, NotificationButtonProps } from "./types";
+export { NotificationButtonType } from "./types";
+export { NOTIFICATION_MODAL_ID } from "./constants";
